LPS-51819 build services

+++ portal-impl/src/com/liferay/portal/service/http/OrganizationServiceSoap.java
@@ -564,7 +564,7 @@ public class OrganizationServiceSoap {
 	attributes for the organization.
 	* @return the organization
 	* @deprecated As of 7.0.0, replaced by {@link #updateOrganization(long,
-	long, String, String, long, long, int, String, boolean,
+	long, String, String, long, long, long, String, boolean,
 	byte[], boolean, List, List, List, List, List,
 	ServiceContext)}
 	*/
