+++ portal-service/src/com/liferay/portal/service/OrganizationLocalServiceWrapper.java
@@ -1446,7 +1446,7 @@ public class OrganizationLocalServiceWrapper implements OrganizationLocalService
 	attributes for the organization.
 	* @return the organization
 	* @deprecated As of 7.0.0, replaced by {@link #updateOrganization(long,
-	long, long, String, String, long, long, int, String, boolean,
+	long, long, String, String, long, long, long, String, boolean,
 	byte[], boolean, ServiceContext)}
 	*/
 	@Deprecated
