+++ portal-service/src/com/liferay/portal/service/OrganizationService.java
@@ -352,7 +352,7 @@ public interface OrganizationService extends BaseService {
 	attributes for the organization.
 	* @return the organization
 	* @deprecated As of 7.0.0, replaced by {@link #updateOrganization(long,
-	long, String, String, long, long, int, String, boolean,
+	long, String, String, long, long, long, String, boolean,
 	byte[], boolean, List, List, List, List, List,
 	ServiceContext)}
 	*/
